:hammer: extract IMAGE_MIME_MAP into shared mime utility
Refs #4294
